플레이어 이동 동기화
remove는 아직 안함. 같은 방향으로 이동시에는 자연스럽게 이동하는데 다른 방향으로 이동할때는 순간이동하는 것이 보임.

+++ Client_GameServerTermProject/Client_GameServerTermProject/Types.h
@@ -177,8 +177,8 @@ struct VectorInt
 		return x * other.y - y * other.x;
 	}
 
-	int32 x = 0;
-	int32 y = 0;
+	int16 x = 0;
+	int16 y = 0;
 };
 
 using Pos = Vector; ;
